helm(e2e): move HTTPRoute-specific setup out of TestMain
The gateway CRD installation, Envoy Gateway deployment, GatewayClass/
Gateway setup, and port-forward are only needed by the HTTPRoute test.

Introduce a suiteSetup hook in main_test.go that each test suite
registers via init(), and move all the gateway logic into a
setupHTTPRoute() function in httproute_test.go. TestMain now only owns
the generic cluster lifecycle and Helm dependency update.
