+++ operations/pyroscope/helm/pyroscope/e2e/main_test.go
@@ -4,13 +4,11 @@ import (
 	"bytes"
 	"fmt"
 	"math/rand"
-	"net/http"
 	"os"
 	"os/exec"
 	"path/filepath"
 	"strings"
 	"testing"
-	"time"
 )
 
 const (
@@ -21,6 +19,10 @@ const (
 // chartDir is the path to the Helm chart root (the e2e/ package directory's parent).
 // Tests are run with the working directory set to the package directory.
 var chartDir = ".."
+
+// suiteSetup is registered by a test suite's init() to run suite-specific
+// infrastructure after the cluster is ready. It returns a cleanup function.
+var suiteSetup func() (func(), error)
 
 func TestMain(m *testing.M) {
 	os.Exit(run(m))
@@ -47,62 +49,7 @@ func run(m *testing.M) (code int) {
 		_ = kind("delete", "cluster", "--name", clusterName)
 	}()
 
-	// ── 2. Gateway API CRDs ──────────────────────────────────────────────────
-	fmt.Println("==> Installing Gateway API CRDs")
-	if err := kubectl("apply", "-f",
-		"https://github.com/kubernetes-sigs/gateway-api/releases/download/"+gwAPIVersion+"/standard-install.yaml",
-	); err != nil {
-		fmt.Fprintf(os.Stderr, "gateway API CRDs: %v\n", err)
-		return 1
-	}
-	for _, crd := range []string{
-		"gateways.gateway.networking.k8s.io",
-		"httproutes.gateway.networking.k8s.io",
-		"gatewayclasses.gateway.networking.k8s.io",
-	} {
-		if err := kubectl("wait", "--for=condition=established", "--timeout=1m", "crd/"+crd); err != nil {
-			fmt.Fprintf(os.Stderr, "wait for CRD %s: %v\n", crd, err)
-			return 1
-		}
-	}
-
-	// ── 3. Envoy Gateway ─────────────────────────────────────────────────────
-	fmt.Println("==> Installing Envoy Gateway", envoyGWVersion)
-	if err := kubectl("apply", "--server-side", "-f",
-		"https://github.com/envoyproxy/gateway/releases/download/"+envoyGWVersion+"/install.yaml",
-	); err != nil {
-		fmt.Fprintf(os.Stderr, "envoy gateway: %v\n", err)
-		return 1
-	}
-	if err := kubectl("wait", "--timeout=5m",
-		"-n", envoyGWNS,
-		"deployment/envoy-gateway",
-		"--for=condition=Available",
-	); err != nil {
-		fmt.Fprintf(os.Stderr, "wait for envoy-gateway deployment: %v\n", err)
-		return 1
-	}
-
-	// ── 4. GatewayClass + Gateway ─────────────────────────────────────────────
-	fmt.Println("==> Applying GatewayClass and Gateway")
-	if err := kubectl("apply", "-f", filepath.Join(chartDir, "ci/gateway-resources.yaml")); err != nil {
-		fmt.Fprintf(os.Stderr, "gateway resources: %v\n", err)
-		return 1
-	}
-	if err := kubectl("wait", "gatewayclass/envoy", "--for=condition=Accepted", "--timeout=60s"); err != nil {
-		fmt.Fprintf(os.Stderr, "wait for GatewayClass: %v\n", err)
-		return 1
-	}
-	if err := kubectl("wait", "gateway/"+gatewayName,
-		"-n", gatewayNS,
-		"--for=condition=Programmed",
-		"--timeout=2m",
-	); err != nil {
-		fmt.Fprintf(os.Stderr, "wait for Gateway: %v\n", err)
-		return 1
-	}
-
-	// ── 5. Helm chart dependencies ────────────────────────────────────────────
+	// ── 2. Helm chart dependencies ────────────────────────────────────────────
 	fmt.Println("==> Updating Helm chart dependencies")
 	_ = helm("repo", "add", "minio", "https://charts.min.io/")
 	_ = helm("repo", "add", "grafana", "https://grafana.github.io/helm-charts")
@@ -111,14 +58,15 @@ func run(m *testing.M) (code int) {
 		return 1
 	}
 
-	// ── 6. Port-forward Envoy proxy ───────────────────────────────────────────
-	fmt.Println("==> Port-forwarding Envoy proxy on :" + gatewayPort)
-	pfCmd, err := startPortForward()
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "port-forward: %v\n", err)
-		return 1
+	// ── 3. Suite-specific setup ───────────────────────────────────────────────
+	if suiteSetup != nil {
+		cleanup, err := suiteSetup()
+		if err != nil {
+			fmt.Fprintf(os.Stderr, "suite setup: %v\n", err)
+			return 1
+		}
+		defer cleanup()
 	}
-	defer pfCmd.Process.Kill() //nolint:errcheck
 
 	return m.Run()
 }
@@ -174,53 +122,6 @@ func runCmd(name string, args ...string) error {
 		return fmt.Errorf("%s %s: %w\n%s", name, strings.Join(args, " "), err, buf.String())
 	}
 	return nil
-}
-
-func startPortForward() (*exec.Cmd, error) {
-	// Find the Envoy proxy service by label.
-	out, err := exec.Command("kubectl",
-		"--context", "kind-"+clusterName,
-		"get", "svc", "-n", envoyGWNS,
-		"-l", fmt.Sprintf(
-			"gateway.envoyproxy.io/owning-gateway-name=%s,gateway.envoyproxy.io/owning-gateway-namespace=%s",
-			gatewayName, gatewayNS,
-		),
-		"-o", "jsonpath={.items[0].metadata.name}",
-	).Output()
-	if err != nil {
-		return nil, fmt.Errorf("get envoy svc: %w", err)
-	}
-	svcName := strings.TrimSpace(string(out))
-	if svcName == "" {
-		return nil, fmt.Errorf("envoy proxy service not found")
-	}
-
-	cmd := exec.Command("kubectl",
-		"--context", "kind-"+clusterName,
-		"port-forward",
-		"-n", envoyGWNS,
-		"svc/"+svcName,
-		gatewayPort+":80",
-	)
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	if err := cmd.Start(); err != nil {
-		return nil, fmt.Errorf("start port-forward: %w", err)
-	}
-
-	// Wait until the port-forward is accepting connections.
-	deadline := time.Now().Add(30 * time.Second)
-	client := &http.Client{Timeout: time.Second}
-	for time.Now().Before(deadline) {
-		resp, err := client.Get("http://localhost:" + gatewayPort + "/")
-		if err == nil {
-			resp.Body.Close()
-			return cmd, nil
-		}
-		time.Sleep(time.Second)
-	}
-	// Return the command even if not ready; the tests will produce clear errors.
-	return cmd, nil
 }
 
 func randStr(n int) string {
